Make some fields nullable in accounts models
The user foreign keys are nullable so that patient and doctor data can
be retained after user account deletion.
The sex field is able to be null so that a default from the two choices
do not have to be forced. otherwise, a default assignment would give
the wrong information (it's better to have no information at all than
to have bad information!)